# storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "School of Net - SON Financeiro API",
        "version": "0.0.1"
    },
    "paths": {
        "/api/login": {
            "post": {
                "summary": "Requisitar token JWT.",
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "properties": {
                                "email": {
                                    "description": "Requisitar token JWT.",
                                    "type": "string"
                                },
                                "password": {
                                    "description": "Requisitar token JWT.",
                                    "type": "string"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Token JWT"
                    }
                }
            }
        },
        "/api/logout": {
            "post": {
                "summary": "Revogar token JWT",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Bearer __token__",
                        "type": "string"
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No content"
                    }
                }
            }
        },
        "/api/refresh_token": {
            "post": {
                "summary": "Renovar token JWT",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Bearer __token__",
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Token JWT"
                    }
                }
            }
        },
...
    },
...
}